refactor climate
